Added confirmation message when deleting history
'history clear' will now display a information message and ask for confirmation.
Closes #11

+++ alpsh/builtins/history.py
@@ -1,6 +1,7 @@
 from alpsh.constants import *
 import os
 import json
+import alpsh.config as config
 
 
 def history(args):
@@ -21,9 +22,20 @@ def history(args):
                 print("SUCCESS : " + data['history'][count]['success'])
                 count += 1
     elif str(args[0] == "clear"):
-        data = {'history': []}
-        with open(LOCATION + HISTORY_FILE, 'w') as history:
-            json.dump(data, history, indent=4)
-        open(LOCATION + HISTORY_FILE_TMP, 'w')
+        print(config.get('text', 'danger')+"-=-=-=-=-=-=-=-WARNING-=-=-=-=-=-=-=\n" +
+              COLORS.CLEAR + "You are about to remove your command history.\n" +
+              "This is generally not necessary and is not recommended to do without reason.\n" +
+              config.get('text', 'danger') + "-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=" + COLORS.CLEAR)
+        response = input("Are you sure you want to continue? [y/N] ")
+        if response == "y":
+            # Remove history
+            data = {'history': []}
+            with open(LOCATION + HISTORY_FILE, 'w') as history:
+                json.dump(data, history, indent=4)
+            open(LOCATION + HISTORY_FILE_TMP, 'w')
+            print(config.get('text', 'warning') + "History cleared!\n" + COLORS.CLEAR)
+        else:
+            # Don't remove history
+            print(config.get('text', 'warning') + "Aborted!\n" + COLORS.CLEAR)
 
     return SHELL_STATUS_RUN
